Adding all the scripts along with the driver
All the forms have been uploaded. automate_extract_gForm is a script that needs to be run on Google Apps Script as it has Google API in use. I have presented the template file for all the questions uploaded as questions_fifth.xlsx.
main script is auto-forms.py which will automate the entire updation of question and answer for Future X Academy.